agents/simple-agent-python-strands/pyproject.toml
[project]
name = "agentcore-simple-agent-python-strands"
version = "0.1.0"
description = "A simple agent for agentcore that uses strands agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=0.1.0",
    "strands-agents>=1.14.0"
]
